[FEAT] 이번주 스터디 코드 추가 1. @RestControllerAdvice를 사용한 에러 핸들링 2. Entity와 DTO 분리 3. @Valid를 이용한 유효성 검증

+++ server/member-api/src/main/java/com/seungh1024/service/MemberService.java
@@ -1,8 +1,15 @@
 package com.seungh1024.service;
 
+import com.seungh1024.dto.MemberDto;
 import com.seungh1024.entity.Member;
+import com.seungh1024.exception.custom.DuplicateMemberException;
 import com.seungh1024.repository.MemberRepository;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.DuplicateKeyException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.util.List;
 
@@ -15,13 +22,20 @@ public class MemberService {
     }
 
     //사용자 등록
-    public boolean createMember(Member member){
-        try{
-            memberRepository.save(member);
-            return true;
-        }catch(Exception e){
-            return false;
+    public Member createMember(MemberDto.MemberJoinRequest memberDto){
+        int memberCheck = memberRepository.countMemberByMemberEmail(memberDto.getMemberEmail());
+        if(memberCheck > 0){ // 1이상이면 중복되는 누군가 있는 것
+
+            throw new DuplicateMemberException(memberDto.getMemberEmail());
         }
+
+        Member member = Member.builder()
+                .memberEmail(memberDto.getMemberEmail())
+                .memberPassword(memberDto.getMemberPassword())
+                .memberName(memberDto.getMemberName())
+                .build();
+        member = memberRepository.save(member);
+        return member;
     }
 
     //전체 사용자 조회
@@ -30,31 +44,32 @@ public class MemberService {
     }
 
     //사용자 이메일로 검색
-    public Member findMemberByEmail(String email){
-        return memberRepository.findMemberByMemberEmail(email);
+   public Member findMemberByEmail(String email){
+       Member member = memberRepository.findMemberByMemberEmail(email);
+       System.out.println("member: "+member);
+       if(member == null){
+           throw new EntityNotFoundException();
+       }
+       return  member;
     }
 
+    //TODO 업데이트,삭제 로그인 연동되면 그걸로 pk받아와서 수정 및 삭제로 변경
     //사용자 비밀번호 업데이트
-    public boolean updateMemberPassword(int pk,String password){
-        try{
-            Member member = memberRepository.findMemberByMemberId(pk);
-            member.updatePassword(password);
-            memberRepository.save(member);
-            return true;
-        }catch(Exception e){
-//            e.printStackTrace();
-            return false;
+    public void updateMemberPassword(int pk,String password){
+        Member member = memberRepository.findMemberByMemberId(pk);
+        if(member == null){
+            throw new EntityNotFoundException();
         }
+        member.updatePassword(password);
+        member = memberRepository.save(member);
     }
 
     //사용자 삭제
-    public boolean deleteMember(int pk){
-        try{
-            Member member = memberRepository.findMemberByMemberId(pk);
-            memberRepository.delete(member);
-            return true;
-        }catch(Exception e){
-            return false;
+    public void deleteMember(int pk){
+        Member member = memberRepository.findMemberByMemberId(pk);
+        if(member == null){
+            throw new EntityNotFoundException();
         }
+        memberRepository.delete(member);
     }
 }
